Update README.md
I changed the README.md file to explain why I loved the series inside the file so much and why they are so unique to me. I have also explained and showcased what my favorite characters are in Naruto, Dragon Ball, and HunterxHunter, and why the series is so treasured by me!